+++ packages/gatsby-plugin-page-creator/src/create-pages-from-collection-builder.ts
@@ -2,6 +2,7 @@
 import { Actions, CreatePagesArgs } from "gatsby"
 import { createPath } from "gatsby-page-utils"
 import { Reporter } from "gatsby"
+import { Options as ISlugifyOptions } from "@sindresorhus/slugify"
 import { reverseLookupParams } from "./extract-query"
 import { getMatchPath } from "./get-match-path"
 import { getCollectionRouteParams } from "./get-collection-route-params"
@@ -11,13 +12,13 @@ import { collectionExtractQueryString } from "./collection-extract-query-string"
 import { isValidCollectionPathImplementation } from "./is-valid-collection-path-implementation"
 import { CODES, prefixId } from "./error-utils"
 
-// TODO: Do we need the ignore argument?
 export async function createPagesFromCollectionBuilder(
   filePath: string,
   absolutePath: string,
   actions: Actions,
   graphql: CreatePagesArgs["graphql"],
-  reporter: Reporter
+  reporter: Reporter,
+  slugifyOptions?: ISlugifyOptions
 ): Promise<void> {
   if (isValidCollectionPathImplementation(absolutePath, reporter) === false) {
     watchCollectionBuilder(absolutePath, ``, [], actions, reporter, () =>
@@ -26,7 +27,8 @@ export async function createPagesFromCollectionBuilder(
         absolutePath,
         actions,
         graphql,
-        reporter
+        reporter,
+        slugifyOptions
       )
     )
     return
@@ -43,7 +45,8 @@ export async function createPagesFromCollectionBuilder(
         absolutePath,
         actions,
         graphql,
-        reporter
+        reporter,
+        slugifyOptions
       )
     )
     return
@@ -78,7 +81,8 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
           absolutePath,
           actions,
           graphql,
-          reporter
+          reporter,
+          slugifyOptions
         )
     )
 
@@ -105,7 +109,12 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
   //    the watcher will use this data to delete the pages if the query changes significantly.
   const paths = nodes.map((node: Record<string, object>) => {
     // URL path for the component and node
-    const { derivedPath, errors } = derivePath(filePath, node, reporter)
+    const { derivedPath, errors } = derivePath(
+      filePath,
+      node,
+      reporter,
+      slugifyOptions
+    )
     const path = createPath(derivedPath)
     // Params is supplied to the FE component on props.params
     const params = getCollectionRouteParams(createPath(filePath), path)
@@ -150,7 +159,8 @@ ${errors.map(error => error.message).join(`\n`)}`.trim(),
         absolutePath,
         actions,
         graphql,
-        reporter
+        reporter,
+        slugifyOptions
       )
   )
 }