+++ packages/gatsby/src/services/initialize.ts
@@ -65,20 +65,14 @@ if (
   sampleSiteForExperiment(`DEV_SSR`, 5)
 ) {
   showExperimentNoticeAfterTimeout(
-    `devSSR`,
-    `
-Your dev experience is about to get better, faster, and stronger!
+    `Server Side Rendering (SSR) in Development`,
+    `gatsby.dev/dev-ssr-feedback`,
+    `which helps surface issues with build errors more quickly. Here's how to try it:
 
-We'll soon be shipping support for SSR in development.
-
-This will help the dev environment more closely mimic builds so you'll catch build errors earlier and fix them faster.
-
-Try out develop SSR *today* by running your site with it enabled:
-
-GATSBY_EXPERIMENTAL_DEV_SSR=true gatsby develop
-
-Please let us know how it goes good, bad, or otherwise at gatsby.dev/dev-ssr-feedback
-      `,
+module.exports = {
+  flags : { DEV_SSR: true },
+  plugins: [...]
+}`,
     1 // Show this immediately to the subset of sites selected.
   )
 }
@@ -181,10 +175,10 @@ export async function initialize({
   }
 
   // Setup flags
-  if (config && config.flags) {
+  if (config) {
     // TODO: this should be handled in FAST_REFRESH configuration and not be one-off here.
     if (
-      config.flags.FAST_REFRESH &&
+      config.flags?.FAST_REFRESH &&
       process.env.GATSBY_HOT_LOADER &&
       process.env.GATSBY_HOT_LOADER !== `fast-refresh`
     ) {
@@ -192,13 +186,14 @@ export async function initialize({
       reporter.warn(
         reporter.stripIndent(`
           Both FAST_REFRESH gatsby-config flag and GATSBY_HOT_LOADER environment variable is used with conflicting setting ("${process.env.GATSBY_HOT_LOADER}").
-          
+
           Will use react-hot-loader.
-          
+
           To use Fast Refresh either do not use GATSBY_HOT_LOADER environment variable or set it to "fast-refresh".
         `)
       )
     }
+
     const availableFlags = require(`../utils/flags`).default
     // Get flags
     const { enabledConfigFlags, unknownFlagMessage, message } = handleFlags(
@@ -226,12 +221,24 @@ export async function initialize({
     })
 
     // Track the usage of config.flags
-    if (enabledConfigFlags.length > 0) {
+    if (config.flags) {
       telemetry.trackFeatureIsUsed(`ConfigFlags`)
     }
   }
 
   process.env.GATSBY_HOT_LOADER = getReactHotLoaderStrategy()
+
+  // TODO: figure out proper way of disabling loading indicator
+  // for now GATSBY_QUERY_ON_DEMAND_LOADING_INDICATOR=false gatsby develop
+  // will work, but we don't want to force users into using env vars
+  if (
+    process.env.GATSBY_EXPERIMENTAL_QUERY_ON_DEMAND &&
+    !process.env.GATSBY_QUERY_ON_DEMAND_LOADING_INDICATOR
+  ) {
+    // if query on demand is enabled and GATSBY_QUERY_ON_DEMAND_LOADING_INDICATOR was not set at all
+    // enable loading indicator
+    process.env.GATSBY_QUERY_ON_DEMAND_LOADING_INDICATOR = `true`
+  }
 
   // theme gatsby configs can be functions or objects
   if (config && config.__experimentalThemes) {