+++ packages/gatsby-source-mongodb/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.8.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-mongodb@2.7.0-next.0...gatsby-source-mongodb@2.8.0-next.0) (2020-12-10)
+
+**Note:** Version bump only for package gatsby-source-mongodb
+
 # [2.7.0-next.0](https://github.com/gatsbyjs/gatsby/compare/gatsby-source-mongodb@2.6.0-next.0...gatsby-source-mongodb@2.7.0-next.0) (2020-11-26)
 
 **Note:** Version bump only for package gatsby-source-mongodb